pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "glyphser"
version = "0.2.0"
description = "Deterministic evidence engine for reproducible ML runtime verification."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Astrocytech" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "cbor2>=5.6.0",
    "PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/Astrocytech/Glyphser"
Repository = "https://github.com/Astrocytech/Glyphser"
Issues = "https://github.com/Astrocytech/Glyphser/issues"

[project.scripts]
glyphser = "glyphser.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.10.0",
    "bandit>=1.7.9",
    "defusedxml>=0.7.1",
    "pip-audit>=2.7.3",
    "pre-commit>=3.7.0",
    "ruff>=0.3.0",
    "build>=1.2.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["glyphser*", "runtime*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []
